move session player to components folder
