+++ src/components/SessionPlayer/SelectorHighlights.tsx
