+++ src/components/SessionPlayer/Timeline/TimelineSlider.tsx
