+++ src/components/SessionPlayer/Viewport.tsx
